dashboard-front/ML_Service/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow==2.19.0
tensorflow_hub
opencv-python
python-multipart
